refactor: split flow models into package
